+++ src/App.css
@@ -22,10 +22,10 @@ body {
   -webkit-text-size-adjust: 100%;
 }
 
-a,
+/* a,
 a:visited {
-  color: #00d8ff
-}
+  color: black
+} */
 
 nav > ul {
   margin: 0;
