+++ src/components/NewReview/NewReview.module.css
@@ -1,29 +1,6 @@
-.profilePage {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.profileName {
-  font-size: 50px;
-  color: #fcd259;
-}
-
-.avatar {
-  max-width: 300px;
-  max-height: 300px;
-  border-radius: 10px;
-}
-
-.reviews {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
 .reviewForm {
   display: flex;
+  margin-top: 20px;
   gap: 10px;
 }
