+++ src/components/Reviews/Reviews.module.css
@@ -1,0 +1,5 @@
+.reviews {
+  display: flex;
+  flex-direction: column;
+  margin-top: 15px;
+}
